portal/requirements.txt
aniso8601==3.0.2
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.4
django-decouple==2.1
django-sass-processor==0.7.2
google-api-core==1.6.0
google-auth==1.6.1
google-cloud-bigquery==1.7.0
google-cloud-core==0.28.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
hamlpy==0.82.2
idna==2.7
libsass==0.16.1
Markdown==3.0.1
oauthlib==2.1.0
Pillow==5.3.0
promise==2.2.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.3.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
ruamel.yaml==0.15.80
Rx==1.6.1
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
urllib3==1.24.1
